:adhesive_bandage: Chore: renamed core function
