+++ web/src/components/core/TablePenilaian.vue
